extract OAuth helpers and improve logout logging
